@@ -9,7 +9,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
-import android.widget.Switch
 import android.widget.Toast
 import androidx.core.content.pm.PackageInfoCompat
 import androidx.fragment.app.Fragment
@@ -48,35 +47,9 @@ import de.rki.coronawarnapp.storage.ExposureSummaryRepository
 import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.storage.tracing.TracingIntervalRepository
 import de.rki.coronawarnapp.transaction.RiskLevelTransaction
-import de.rki.coronawarnapp.ui.viewLifecycle
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.KeyFileHelper
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_enter_other_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_get_check_exposure
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_get_exposure_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_scan_qr_code
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_share_my_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_submit_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_api_test_start
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_calculate_risk_level
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_clear_db
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_insert_exposure_summary
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_retrieve_exposure_summary
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_tracing_duration_in_retention_period
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.button_tracing_intervals
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_exposure_summary_attenuation
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_exposure_summary_daysSinceLastExposure
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_exposure_summary_matchedKeyCount
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_exposure_summary_maximumRiskScore
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_exposure_summary_summationRiskScore
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_googlePlayServices_version
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_latest_key_date
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.label_my_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.qr_code_viewpager
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.test_api_switch_last_three_hours_from_server
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.test_api_switch_background_notifications
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.text_my_keys
-import kotlinx.android.synthetic.deviceForTesters.fragment_test_for_a_p_i.text_scanned_key
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -88,7 +61,8 @@ import java.lang.reflect.Type
 import java.util.UUID
 
 @SuppressWarnings("TooManyFunctions", "MagicNumber", "LongMethod")
-class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHelper.Callback {
+class TestForAPIFragment : Fragment(R.layout.fragment_test_for_a_p_i),
+    InternalExposureNotificationPermissionHelper.Callback {
 
     companion object {
         const val CONFIG_SCORE = 8
@@ -119,16 +93,10 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
     private lateinit var qrPagerAdapter: RecyclerView.Adapter<QRPagerAdapter.QRViewHolder>
 
     // Data and View binding
-    private var binding: FragmentTestForAPIBinding by viewLifecycle()
+    private val binding: FragmentTestForAPIBinding by viewBindingLazy()
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-
-        // get the binding reference by inflating it with the current layout
-        binding = FragmentTestForAPIBinding.inflate(inflater)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         // set the viewmmodel variable that will be used for data binding
         binding.tracingViewModel = tracingViewModel
@@ -136,20 +104,14 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
         // set the lifecycleowner for LiveData
         binding.lifecycleOwner = this
 
-        // Inflate the layout for this fragment
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
         val v: Long = PackageInfoCompat.getLongVersionCode(
             activity?.packageManager!!.getPackageInfo(
                 GoogleApiAvailability.GOOGLE_PLAY_SERVICES_PACKAGE,
                 0
             )
         )
-        label_googlePlayServices_version.text = "Google Play Services version: " + v.toString()
+        binding.labelGooglePlayServicesVersion.text =
+            "Google Play Services version: " + v.toString()
 
         token = UUID.randomUUID().toString()
 
@@ -158,19 +120,19 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
 
         getExposureKeys()
 
-        qrPager = qr_code_viewpager
+        qrPager = binding.qrCodeViewpager
         qrPagerAdapter = QRPagerAdapter()
         qrPager.adapter = qrPagerAdapter
 
-        button_api_test_start.setOnClickListener {
+        binding.buttonApiTestStart.setOnClickListener {
             start()
         }
 
-        button_api_get_exposure_keys.setOnClickListener {
+        binding.buttonApiGetExposureKeys.setOnClickListener {
             getExposureKeys()
         }
 
-        val last3HoursSwitch = test_api_switch_last_three_hours_from_server as Switch
+        val last3HoursSwitch = binding.testApiSwitchLastThreeHoursFromServer
         last3HoursSwitch.isChecked = LocalData.last3HoursMode()
         last3HoursSwitch.setOnClickListener {
             val isLast3HoursModeEnabled = last3HoursSwitch.isChecked
@@ -178,7 +140,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             LocalData.last3HoursMode(isLast3HoursModeEnabled)
         }
 
-        val backgroundNotificationSwitch = test_api_switch_background_notifications as Switch
+        val backgroundNotificationSwitch = binding.testApiSwitchBackgroundNotifications
         backgroundNotificationSwitch.isChecked = LocalData.backgroundNotification()
         backgroundNotificationSwitch.setOnClickListener {
             val isBackgroundNotificationsActive = backgroundNotificationSwitch.isChecked
@@ -186,26 +148,26 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             LocalData.backgroundNotification(isBackgroundNotificationsActive)
         }
 
-        button_api_get_check_exposure.setOnClickListener {
+        binding.buttonApiGetCheckExposure.setOnClickListener {
             checkExposure()
         }
 
-        button_api_scan_qr_code.setOnClickListener {
+        binding.buttonApiScanQrCode.setOnClickListener {
             IntentIntegrator.forSupportFragment(this)
                 .setOrientationLocked(false)
                 .setBeepEnabled(false)
                 .initiateScan()
         }
 
-        button_api_share_my_keys.setOnClickListener {
+        binding.buttonApiShareMyKeys.setOnClickListener {
             shareMyKeys()
         }
 
-        button_api_enter_other_keys.setOnClickListener {
+        binding.buttonApiEnterOtherKeys.setOnClickListener {
             enterOtherKeys()
         }
 
-        button_api_submit_keys.setOnClickListener {
+        binding.buttonApiSubmitKeys.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 try {
                     internalExposureNotificationPermissionHelper.requestPermissionToShareKeys()
@@ -220,7 +182,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             }
         }
 
-        button_calculate_risk_level.setOnClickListener {
+        binding.buttonCalculateRiskLevel.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 try {
                     RiskLevelTransaction.start()
@@ -230,7 +192,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             }
         }
 
-        button_insert_exposure_summary.setOnClickListener {
+        binding.buttonInsertExposureSummary.setOnClickListener {
             // Now broadcasts them to the worker.
             val intent = Intent(
                 context,
@@ -240,7 +202,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             context?.sendBroadcast(intent)
         }
 
-        button_retrieve_exposure_summary.setOnClickListener {
+        binding.buttonRetrieveExposureSummary.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 showToast(
                     ExposureSummaryRepository.getExposureSummaryRepository()
@@ -249,7 +211,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             }
         }
 
-        button_clear_db.setOnClickListener {
+        binding.buttonClearDb.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 withContext(Dispatchers.IO) {
                     AppDatabase.getInstance(requireContext()).clearAllTables()
@@ -257,7 +219,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             }
         }
 
-        button_tracing_intervals.setOnClickListener {
+        binding.buttonTracingIntervals.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 showToast(
                     TracingIntervalRepository.getDateRepository(requireContext()).getIntervals()
@@ -266,7 +228,7 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             }
         }
 
-        button_tracing_duration_in_retention_period.setOnClickListener {
+        binding.buttonTracingDurationInRetentionPeriod.setOnClickListener {
             tracingViewModel.viewModelScope.launch {
                 showToast(TimeVariables.getActiveTracingDaysInRetentionPeriod().toString())
             }
@@ -290,8 +252,8 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
     private val onScannedKey = { key: AppleLegacyKeyExchange.Key? ->
         Timber.i("keys scanned..")
         key?.let {
-            text_scanned_key.text = prettyKey(key)
-            text_scanned_key.visibility = View.VISIBLE
+            binding.textScannedKey.text = prettyKey(key)
+            binding.textScannedKey.visibility = View.VISIBLE
 //            text_scanned_key.movementMethod = ScrollingMovementMethod.getInstance()
         }
         otherExposureKeyList.add(key!!)
@@ -407,27 +369,27 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
 
     private fun updateExposureSummaryDisplay(exposureSummary: ExposureSummary?) {
 
-        label_exposure_summary_matchedKeyCount.text = getString(
+        binding.labelExposureSummaryMatchedKeyCount.text = getString(
             R.string.test_api_body_matchedKeyCount,
             (exposureSummary?.matchedKeyCount ?: "-").toString()
         )
 
-        label_exposure_summary_daysSinceLastExposure.text = getString(
+        binding.labelExposureSummaryDaysSinceLastExposure.text = getString(
             R.string.test_api_body_daysSinceLastExposure,
             (exposureSummary?.daysSinceLastExposure ?: "-").toString()
         )
 
-        label_exposure_summary_maximumRiskScore.text = getString(
+        binding.labelExposureSummaryMaximumRiskScore.text = getString(
             R.string.test_api_body_maximumRiskScore,
             (exposureSummary?.maximumRiskScore ?: "-").toString()
         )
 
-        label_exposure_summary_summationRiskScore.text = getString(
+        binding.labelExposureSummarySummationRiskScore.text = getString(
             R.string.test_api_body_summation_risk,
             (exposureSummary?.summationRiskScore ?: "-").toString()
         )
 
-        label_exposure_summary_attenuation.text = getString(
+        binding.labelExposureSummaryAttenuation.text = getString(
             R.string.test_api_body_attenuation,
             (exposureSummary?.attenuationDurationsInMinutes?.joinToString() ?: "-").toString()
         )
@@ -444,15 +406,16 @@ class TestForAPIFragment : Fragment(), InternalExposureNotificationPermissionHel
             R.string.test_api_body_my_keys,
             myKeys?.size ?: 0
         )
-        label_my_keys.text = myKeysLabelAndCount
-        text_my_keys.text = myExposureKeysJSON
+        binding.labelMyKeys.text = myKeysLabelAndCount
+        binding.textMyKeys.text = myExposureKeysJSON
 
-        myKeys?.maxBy { it.rollingStartIntervalNumber }?.rollingStartIntervalNumber?.toLong()?.let {
-            val ms = it * 60L * 10L * 1000L
-            val dateString = DateTime(ms, DateTimeZone.UTC)
+        myKeys?.let { myKeys -> this.maxByOrNull<TemporaryExposureKey, Int>({ it.rollingStartIntervalNumber }) }?.rollingStartIntervalNumber.toLong()
+            .let {
+                val ms = it * 60L * 10L * 1000L
+                val dateString = DateTime(ms, DateTimeZone.UTC)
 
-            label_latest_key_date.text = "Latest key is from: $dateString"
-        }
+                binding.labelLatestKeyDate.text = "Latest key is from: $dateString"
+            }
     }
 
     private fun showToast(message: String) {
