@@ -3,34 +3,20 @@ package de.rki.coronawarnapp.ui.information
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentInformationAboutBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
-import de.rki.coronawarnapp.ui.viewLifecycle
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Basic Fragment which only displays static content.
  */
-class InformationAboutFragment : Fragment() {
-    companion object {
-        private val TAG: String? = InformationAboutFragment::class.simpleName
-    }
+class InformationAboutFragment : Fragment(R.layout.fragment_information_about) {
 
-    private var binding: FragmentInformationAboutBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentInformationAboutBinding.inflate(inflater)
-        return binding.root
-    }
+    private val binding: FragmentInformationAboutBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
