@@ -2,35 +2,21 @@ package de.rki.coronawarnapp.ui.information
 
 import android.os.Bundle
 import android.text.method.LinkMovementMethod
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentInformationLegalBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
 import de.rki.coronawarnapp.util.convertToHyperlink
-import de.rki.coronawarnapp.ui.viewLifecycle
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Basic Fragment which only displays static content.
  */
-class InformationLegalFragment : Fragment() {
-    companion object {
-        private val TAG: String? = InformationLegalFragment::class.simpleName
-    }
+class InformationLegalFragment : Fragment(R.layout.fragment_information_legal) {
 
-    private var binding: FragmentInformationLegalBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentInformationLegalBinding.inflate(inflater)
-        return binding.root
-    }
+    private val binding: FragmentInformationLegalBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
