@@ -3,35 +3,21 @@ package de.rki.coronawarnapp.ui.onboarding
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentOnboardingBinding
 import de.rki.coronawarnapp.ui.doNavigate
-import de.rki.coronawarnapp.ui.viewLifecycle
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * Onboarding starting point.
  */
-class OnboardingFragment : Fragment() {
-    companion object {
-        private val TAG: String? = OnboardingFragment::class.simpleName
-    }
+class OnboardingFragment : Fragment(R.layout.fragment_onboarding) {
 
-    private var binding: FragmentOnboardingBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentOnboardingBinding.inflate(inflater)
-        return binding.root
-    }
+    private val binding: FragmentOnboardingBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -46,10 +32,12 @@ class OnboardingFragment : Fragment() {
     private fun setLinks() {
         binding.onboardingInclude.onboardingEasyLanguage
             .setOnClickListener {
-            val browserIntent = Intent(Intent.ACTION_VIEW,
-                Uri.parse(getString(R.string.onboarding_tracing_easy_language_explanation_url)))
-            startActivity(browserIntent)
-        }
+                val browserIntent = Intent(
+                    Intent.ACTION_VIEW,
+                    Uri.parse(getString(R.string.onboarding_tracing_easy_language_explanation_url))
+                )
+                startActivity(browserIntent)
+            }
     }
 
     override fun onResume() {
