@@ -1,15 +1,14 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.appcompat.app.AlertDialog
 import androidx.core.app.NotificationManagerCompat
 import androidx.fragment.app.Fragment
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentOnboardingNotificationsBinding
-import de.rki.coronawarnapp.ui.viewLifecycle
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This fragment ask the user if he wants to get notifications and finishes the onboarding afterwards.
@@ -17,21 +16,9 @@ import de.rki.coronawarnapp.ui.viewLifecycle
  * @see NotificationManagerCompat
  * @see AlertDialog
  */
-class OnboardingNotificationsFragment : Fragment() {
-    companion object {
-        private val TAG: String? = OnboardingNotificationsFragment::class.simpleName
-    }
+class OnboardingNotificationsFragment : Fragment(R.layout.fragment_onboarding_notifications) {
 
-    private var binding: FragmentOnboardingNotificationsBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentOnboardingNotificationsBinding.inflate(inflater)
-        return binding.root
-    }
+    private val binding: FragmentOnboardingNotificationsBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
