@@ -1,34 +1,21 @@
 package de.rki.coronawarnapp.ui.onboarding
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentOnboardingPrivacyBinding
 import de.rki.coronawarnapp.ui.doNavigate
-import de.rki.coronawarnapp.ui.viewLifecycle
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This fragment informs the user regarding privacy.
  */
-class OnboardingPrivacyFragment : Fragment() {
-    companion object {
-        private val TAG: String? = OnboardingPrivacyFragment::class.simpleName
-    }
+class OnboardingPrivacyFragment : Fragment(R.layout.fragment_onboarding_privacy) {
 
-    private var binding: FragmentOnboardingPrivacyBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentOnboardingPrivacyBinding.inflate(inflater)
-        return binding.root
-    }
+    private val binding: FragmentOnboardingPrivacyBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
