@@ -2,9 +2,7 @@ package de.rki.coronawarnapp.ui.onboarding
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
@@ -18,8 +16,8 @@ import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationPermissionHelper
 import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.ui.doNavigate
-import de.rki.coronawarnapp.ui.viewLifecycle
 import de.rki.coronawarnapp.util.DialogHelper
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 import kotlinx.coroutines.launch
 
 /**
@@ -28,7 +26,7 @@ import kotlinx.coroutines.launch
  * @see InternalExposureNotificationPermissionHelper
  * @see AlertDialog
  */
-class OnboardingTracingFragment : Fragment(),
+class OnboardingTracingFragment : Fragment(R.layout.fragment_onboarding_tracing),
     InternalExposureNotificationPermissionHelper.Callback {
 
     companion object {
@@ -36,7 +34,7 @@ class OnboardingTracingFragment : Fragment(),
     }
 
     private lateinit var internalExposureNotificationPermissionHelper: InternalExposureNotificationPermissionHelper
-    private var binding: FragmentOnboardingTracingBinding by viewLifecycle()
+    private val binding: FragmentOnboardingTracingBinding by viewBindingLazy()
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         internalExposureNotificationPermissionHelper.onResolutionComplete(
@@ -49,15 +47,6 @@ class OnboardingTracingFragment : Fragment(),
         super.onCreate(savedInstanceState)
         internalExposureNotificationPermissionHelper =
             InternalExposureNotificationPermissionHelper(this, this)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentOnboardingTracingBinding.inflate(inflater)
-        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -123,7 +112,7 @@ class OnboardingTracingFragment : Fragment(),
             } catch (exception: Exception) {
                 exception.report(
                     ExceptionCategory.EXPOSURENOTIFICATION,
-                    OnboardingTracingFragment.TAG,
+                    TAG,
                     null
                 )
             }
