@@ -1,18 +1,17 @@
 package de.rki.coronawarnapp.ui.settings
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSettingsBackgroundPriorityBinding
 import de.rki.coronawarnapp.ui.base.startActivitySafely
 import de.rki.coronawarnapp.ui.main.MainActivity
-import de.rki.coronawarnapp.ui.viewLifecycle
 import de.rki.coronawarnapp.ui.viewmodel.SettingsViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This is the setting background priority page. Here the user sees the background priority setting status.
@@ -21,27 +20,16 @@ import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
  * @see TracingViewModel
  * @see SettingsViewModel
  */
-class SettingsBackgroundPriorityFragment : Fragment() {
-    companion object {
-        private val TAG: String? = SettingsBackgroundPriorityFragment::class.simpleName
-    }
+class SettingsBackgroundPriorityFragment :
+    Fragment(R.layout.fragment_settings_background_priority) {
 
     private val settingsViewModel: SettingsViewModel by activityViewModels()
-    private var binding: FragmentSettingsBackgroundPriorityBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentSettingsBackgroundPriorityBinding.inflate(inflater)
-        binding.settingsViewModel = settingsViewModel
-        binding.lifecycleOwner = this
-        return binding.root
-    }
+    private val binding: FragmentSettingsBackgroundPriorityBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.settingsViewModel = settingsViewModel
+        binding.lifecycleOwner = this
         setButtonOnClickListener()
     }
 
