@@ -1,19 +1,18 @@
 package de.rki.coronawarnapp.ui.settings
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FragmentSettingsNotificationsBinding
 import de.rki.coronawarnapp.ui.main.MainActivity
-import de.rki.coronawarnapp.ui.viewLifecycle
 import de.rki.coronawarnapp.ui.viewmodel.SettingsViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.ExternalActionHelper
 import de.rki.coronawarnapp.util.IGNORE_CHANGE_TAG
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
 /**
  * This is the setting notification page. Here the user sees his os notifications settings status.
@@ -23,27 +22,15 @@ import de.rki.coronawarnapp.util.IGNORE_CHANGE_TAG
  * @see TracingViewModel
  * @see SettingsViewModel
  */
-class SettingsNotificationFragment : Fragment() {
-    companion object {
-        private val TAG: String? = SettingsNotificationFragment::class.simpleName
-    }
+class SettingsNotificationFragment : Fragment(R.layout.fragment_settings_notifications) {
 
     private val settingsViewModel: SettingsViewModel by activityViewModels()
-    private var binding: FragmentSettingsNotificationsBinding by viewLifecycle()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentSettingsNotificationsBinding.inflate(inflater)
-        binding.settingsViewModel = settingsViewModel
-        binding.lifecycleOwner = this
-        return binding.root
-    }
+    private val binding: FragmentSettingsNotificationsBinding by viewBindingLazy()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        binding.settingsViewModel = settingsViewModel
+        binding.lifecycleOwner = this
         setButtonOnClickListener()
     }
 
