@@ -2,9 +2,7 @@ package de.rki.coronawarnapp.ui.submission
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.view.accessibility.AccessibilityEvent
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
@@ -20,23 +18,20 @@ import de.rki.coronawarnapp.exception.http.CwaServerError
 import de.rki.coronawarnapp.exception.http.ForbiddenException
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationPermissionHelper
 import de.rki.coronawarnapp.ui.doNavigate
-import de.rki.coronawarnapp.ui.viewLifecycle
 import de.rki.coronawarnapp.ui.viewmodel.SubmissionViewModel
 import de.rki.coronawarnapp.ui.viewmodel.TracingViewModel
 import de.rki.coronawarnapp.util.DialogHelper
 import de.rki.coronawarnapp.util.observeEvent
+import de.rki.coronawarnapp.util.ui.viewBindingLazy
 
-class SubmissionResultPositiveOtherWarningFragment : Fragment(),
+class SubmissionResultPositiveOtherWarningFragment :
+    Fragment(R.layout.fragment_submission_positive_other_warning),
     InternalExposureNotificationPermissionHelper.Callback {
-
-    companion object {
-        private val TAG: String? = SubmissionResultPositiveOtherWarningFragment::class.simpleName
-    }
 
     private val submissionViewModel: SubmissionViewModel by activityViewModels()
     private val tracingViewModel: TracingViewModel by activityViewModels()
 
-    private var binding: FragmentSubmissionPositiveOtherWarningBinding by viewLifecycle()
+    private val binding: FragmentSubmissionPositiveOtherWarningBinding by viewBindingLazy()
     private lateinit var internalExposureNotificationPermissionHelper:
             InternalExposureNotificationPermissionHelper
 
@@ -52,20 +47,6 @@ class SubmissionResultPositiveOtherWarningFragment : Fragment(),
         super.onResume()
         binding.submissionPositiveOtherPrivacyContainer.sendAccessibilityEvent(AccessibilityEvent.TYPE_ANNOUNCEMENT)
         tracingViewModel.refreshIsTracingEnabled()
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        internalExposureNotificationPermissionHelper =
-            InternalExposureNotificationPermissionHelper(this, this)
-        binding = FragmentSubmissionPositiveOtherWarningBinding.inflate(inflater)
-        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, backCallback)
-        binding.submissionViewModel = submissionViewModel
-        binding.lifecycleOwner = this
-        return binding.root
     }
 
     private fun buildErrorDialog(exception: Exception): DialogHelper.DialogInstance {
@@ -126,6 +107,12 @@ class SubmissionResultPositiveOtherWarningFragment : Fragment(),
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        internalExposureNotificationPermissionHelper =
+            InternalExposureNotificationPermissionHelper(this, this)
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, backCallback)
+        binding.submissionViewModel = submissionViewModel
+        binding.lifecycleOwner = this
+
         setButtonOnClickListener()
 
         submissionViewModel.submissionError.observeEvent(viewLifecycleOwner) {
@@ -149,21 +136,23 @@ class SubmissionResultPositiveOtherWarningFragment : Fragment(),
     }
 
     /**
-    * Opens a Dialog that warns user
-    * when they're about to cancel the submission flow
-    * @see DialogHelper
-    * @see navigateToSubmissionResultFragment
-    */
+     * Opens a Dialog that warns user
+     * when they're about to cancel the submission flow
+     * @see DialogHelper
+     * @see navigateToSubmissionResultFragment
+     */
     fun handleSubmissionCancellation() {
-        DialogHelper.showDialog(DialogHelper.DialogInstance(
-            requireActivity(),
-            R.string.submission_error_dialog_confirm_cancellation_title,
-            R.string.submission_error_dialog_confirm_cancellation_body,
-            R.string.submission_error_dialog_confirm_cancellation_button_positive,
-            R.string.submission_error_dialog_confirm_cancellation_button_negative,
-            true,
-            ::navigateToSubmissionResultFragment
-        ))
+        DialogHelper.showDialog(
+            DialogHelper.DialogInstance(
+                requireActivity(),
+                R.string.submission_error_dialog_confirm_cancellation_title,
+                R.string.submission_error_dialog_confirm_cancellation_body,
+                R.string.submission_error_dialog_confirm_cancellation_button_positive,
+                R.string.submission_error_dialog_confirm_cancellation_button_negative,
+                true,
+                ::navigateToSubmissionResultFragment
+            )
+        )
     }
 
     private fun navigateToSubmissionResultFragment() =
